# public/openapi/moim-api.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "MOIM API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://{env}.moim.id/api/v1",
            "variables": {
                "env": {
                    "enum": [
                        "dev",
                        "www"
                    ],
                    "default": "dev"
                }
            }
        }
    ],
...
    "components": {
        "schemas": {
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "messages": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user logged in data",
                "description": "",
                "parameters": [
                    {
                        "name": "app-token",
                        "in": "header",
                        "description": "Fireauth token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "keyword",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return user logged in data successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "messages": {},
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "fullname": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "username": {
                                                    "type": "string"
                                                },
                                                "gender": {
                                                    "type": "string"
                                                },
                                                "account": {
                                                    "type": "string"
                                                },
                                                "date_birth": {
                                                    "type": "string"
                                                },
                                                "bio": {
                                                    "type": "string"
                                                },
                                                "img_profile": {
                                                    "type": "string"
                                                },
                                                "phone_no": {
                                                    "type": "string"
                                                },
                                                "verified_phone_no": {
                                                    "type": "string"
                                                },
                                                "followers": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "following": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "moim": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "preferences": {
                                                    "type": "object"
                                                },
                                                "categories": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "created_by": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                },
                                                "updated_by": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                },
                                                "deleted_by": {
                                                    "type": "string"
                                                },
                                                "deleted_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                },
                                                "gander": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register User",
                "description": "",
                "parameters": [
                    {
                        "name": "app-token",
                        "in": "header",
                        "description": "Fireauth token",
                        "required": true,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "description": "Data User",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "date_birth": {
                                        "type": "string",
                                        "format": "date-time",
                                        "description": "date in UTC"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "fullname": {
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        },
                                                        "username": {
                                                            "type": "string"
                                                        },
                                                        "gender": {
                                                            "type": "string"
                                                        },
                                                        "account": {
                                                            "type": "string"
                                                        },
                                                        "date_birth": {
                                                            "type": "string"
                                                        },
                                                        "bio": {
                                                            "type": "string"
                                                        },
                                                        "img_profile": {
                                                            "type": "string"
                                                        },
                                                        "phone_no": {
                                                            "type": "string"
                                                        },
                                                        "verified_phone_no": {
                                                            "type": "string"
                                                        },
                                                        "followers": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "following": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "moim": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "preferences": {
                                                            "type": "object"
                                                        },
                                                        "categories": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "created_by": {
                                                            "type": "string"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "datetime"
                                                        },
                                                        "updated_by": {
                                                            "type": "string"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "datetime"
                                                        },
                                                        "deleted_by": {
                                                            "type": "string"
                                                        },
                                                        "deleted_at": {
                                                            "type": "string",
                                                            "format": "datetime"
                                                        },
                                                        "gander": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users?key":{
            "get":{
                "tags": [
                    "User"
                ],
                "summary": "pencarian berdasarkan keyword",
                "description": "",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "keyword",
                        "required": true,
                        "schema":{}
                    },
                    {
                        "name": "app-token",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema":{}
                    }
                ],
                "responses": {
                    "200":{
                        "description": "",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data":{
                                                    "type": "array",
                                                    "items":{
                                                        "properties": {
                                                            "user_id":{
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}